requirements.txt
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-heroku==0.3.1
django-pwa==1.0.8
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djoser==2.0.3
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
model-bakery==1.1.0
oauthlib==3.1.0
openapi-codec==1.3.2
pathspec==0.7.0
pbr==5.4.5
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.0.1
zipp==3.1.0
